backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
psutil==5.9.8
python-dotenv==1.0.1
docker==7.1.0
